+++ legadilo/feeds/admin.py
@@ -11,7 +11,6 @@ from legadilo.feeds.models import (
     Feed,
     FeedArticle,
     FeedCategory,
-    FeedDeletedArticle,
     FeedTag,
     FeedUpdate,
 )
@@ -88,8 +87,3 @@ class FeedCategoryAdmin(admin.ModelAdmin):
 class FeedArticleAdmin(admin.ModelAdmin):
     search_fields = ["feed__title", "article__title", "article__url"]
     autocomplete_fields = ["feed", "article"]
-
-
-@admin.register(FeedDeletedArticle)
-class FeedDeletedArticleAdmin(admin.ModelAdmin):
-    pass