+++ legadilo/feeds/tests/test_commands/test_clean_data.py
@@ -6,7 +6,7 @@ import pytest
 import time_machine
 from django.core.management import call_command
 
-from legadilo.feeds.models import FeedDeletedArticle, FeedUpdate
+from legadilo.feeds.models import FeedUpdate
 from legadilo.feeds.tests.factories import FeedArticleFactory, FeedFactory, FeedUpdateFactory
 from legadilo.reading.models import Article, ArticleFetchError
 from legadilo.reading.tests.factories import ArticleFactory, ArticleFetchErrorFactory
@@ -48,12 +48,18 @@ class TestCleanDataCommand:
         article_linked_to_forever_feed = ArticleFactory(
             title="Manually added", user=user, read_at=utcdt(2024, 6, 1)
         )
-        FeedArticleFactory(feed=feed_keep_article_forever, article=article_linked_to_forever_feed)
+        feed_article_to_keep = FeedArticleFactory(
+            feed=feed_keep_article_forever, article=article_linked_to_forever_feed
+        )
         feed_to_cleanup = FeedFactory(user=user, article_retention_time=1)
         article_linked_to_feed_to_cleanup = ArticleFactory(user=user, read_at=utcdt(2024, 6, 1))
-        FeedArticleFactory(feed=feed_to_cleanup, article=article_linked_to_feed_to_cleanup)
+        feed_article_to_cleanup = FeedArticleFactory(
+            feed=feed_to_cleanup, article=article_linked_to_feed_to_cleanup
+        )
         unread_article_linked_to_feed_to_cleanup = ArticleFactory(user=user)
-        FeedArticleFactory(feed=feed_to_cleanup, article=unread_article_linked_to_feed_to_cleanup)
+        feed_article_unread = FeedArticleFactory(
+            feed=feed_to_cleanup, article=unread_article_linked_to_feed_to_cleanup
+        )
 
         with time_machine.travel("2024-07-01 00:00:00"):
             call_command("clean_data")
@@ -63,4 +69,9 @@ class TestCleanDataCommand:
             article_linked_to_forever_feed,
             unread_article_linked_to_feed_to_cleanup,
         ]
-        assert FeedDeletedArticle.objects.count() > 0
+        feed_article_to_keep.refresh_from_db()
+        assert feed_article_to_keep.article == article_linked_to_forever_feed
+        feed_article_to_cleanup.refresh_from_db()
+        assert feed_article_to_cleanup.article is None
+        feed_article_unread.refresh_from_db()
+        assert feed_article_unread.article == unread_article_linked_to_feed_to_cleanup