+++ legadilo/reading/api.py
@@ -25,7 +25,6 @@ from legadilo.reading.services.article_fetching import (
     build_article_data_from_content,
     get_article_from_url,
 )
-from legadilo.reading.services.delete_article import delete_article
 from legadilo.users.models import User
 from legadilo.users.user_types import AuthenticatedApiRequest
 from legadilo.utils.api import ApiError, NotSet, update_model_from_schema
@@ -209,7 +208,7 @@ def _update_article_tags(user: User, article: Article, new_tags: tuple[str, ...]
 def delete_article_view(request: AuthenticatedApiRequest, article_id: int):
     article = get_object_or_404(Article, id=article_id, user=request.auth)
 
-    delete_article(article)
+    article.delete()
 
     return HTTPStatus.NO_CONTENT, None